need to add web + dandy, leaderboard and tooltips

+++ Client/Assets/Assets.hh
@@ -21,4 +21,4 @@ void draw_static_petal_single(uint8_t, Renderer &);
 
 void draw_static_petal(uint8_t, Renderer &);
 
-void draw_static_mob(uint8_t, Renderer &, float, float, uint8_t);
+void draw_static_mob(uint8_t, Renderer &, float, float, uint32_t);
